Initial commit
This is the initial commit for the project.  At this time, the
application is a bit under-tested and under-specified, but it does
have the ability to accept the contents of a GPX file and return
either a re-direct to a Google Maps image of it, or an elevation
augmented version of the same file.